Changed `Area3D` to always do back-face collisions
