+++ src/shapes/jolt_concave_polygon_shape_impl_3d.hpp
@@ -21,8 +21,6 @@ public:
 private:
 	JPH::ShapeRefC _build() const override;
 
-	JPH::ShapeRefC _build_double_sided(const JPH::Shape* p_shape) const;
-
 	PackedVector3Array faces;
 
 	bool backface_collision = false;
