+++ src/shapes/jolt_shape_impl_3d.cpp
@@ -1,6 +1,7 @@
 #include "jolt_shape_impl_3d.hpp"
 
 #include "objects/jolt_shaped_object_impl_3d.hpp"
+#include "shapes/jolt_custom_double_sided_shape.hpp"
 #include "shapes/jolt_custom_user_data_shape.hpp"
 
 namespace {
@@ -158,6 +159,24 @@ JPH::ShapeRefC JoltShapeImpl3D::with_user_data(const JPH::Shape* p_shape, uint64
 		shape_result.HasError(),
 		vformat(
 			"Failed to override user data. "
+			"It returned the following error: '%s'.",
+			to_godot(shape_result.GetError())
+		)
+	);
+
+	return shape_result.Get();
+}
+
+JPH::ShapeRefC JoltShapeImpl3D::with_double_sided(const JPH::Shape* p_shape) {
+	ERR_FAIL_NULL_D(p_shape);
+
+	const JoltCustomDoubleSidedShapeSettings shape_settings(p_shape);
+	const JPH::ShapeSettings::ShapeResult shape_result = shape_settings.Create();
+
+	ERR_FAIL_COND_D_MSG(
+		shape_result.HasError(),
+		vformat(
+			"Failed to make shape double-sided. "
 			"It returned the following error: '%s'.",
 			to_godot(shape_result.GetError())
 		)
